PEER-248: Move files
