+++ frontend/src/routes/questions/table-columns.tsx
